+++ app/src/main/java/com/blityapps/blitynews/MainActivity.kt
@@ -3,8 +3,6 @@ package com.blityapps.blitynews
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.absoluteOffset
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
@@ -12,7 +10,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import com.blityapps.blitynews.Theming.BlityNewsTheme
+import com.blityapps.blitynews.theming.BlityNewsTheme
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -24,7 +22,7 @@ class MainActivity : AppCompatActivity() {
 
     @Composable
     fun Main() {
-        BlityNewsTheme {
+        BlityNewsTheme.Theme {
             Surface(modifier = Modifier.fillMaxWidth()) {
                 Text(text = stringResource(id = R.string.app_name))
             }
@@ -34,7 +32,7 @@ class MainActivity : AppCompatActivity() {
     @Preview
     @Composable
     fun MainPreviewLight() {
-        BlityNewsTheme(isDark = false) {
+        BlityNewsTheme.Theme(isDark = false) {
             Surface(modifier = Modifier.fillMaxWidth()) {
                 Text(text = stringResource(id = R.string.app_name))
             }
@@ -44,7 +42,7 @@ class MainActivity : AppCompatActivity() {
     @Preview
     @Composable
     fun MainPreviewDark() {
-        BlityNewsTheme(isDark = true) {
+        BlityNewsTheme.Theme(isDark = true) {
             Surface(modifier = Modifier.fillMaxWidth()) {
                 Text(text = stringResource(id = R.string.app_name))
             }
